+++ src/Input.cpp
@@ -27,7 +27,7 @@ int Input::applyAnalysisWindow(const Eigen::Ref<const Eigen::ArrayXXf> &input)
 	const auto half = analysisWindowBasic.rows() / 2;
 	Window::Apply::special<false>(analysisWindowBasic.head(half), input.bottomRows(input.rows() / 2).topRows(half), windowedInput.topRows(half));
 	Window::Apply::special<false>(analysisWindowBasic.tail(half), input.topRows(input.rows() / 2).bottomRows(half), windowedInput.bottomRows(half));
-	return Bungee::log2(windowedInput.rows());
+	return Bungee::log2((int)windowedInput.rows());
 }
 
 } // namespace Bungee
